- agregado en la ruta de niño el qr
