+++ app/src/main/java/com/example/comusenias/presentation/ui/theme/Size.kt
@@ -41,8 +41,9 @@ const val size150: Int = 150
 const val SIZE150: Int = 150
 const val SIZE160: Int = 160
 const val SIZE220: Int = 220
-const val SIZE300: Int = 300
 const val SIZE250: Int = 250
+const val SIZE300: Int = 300
+const val SIZE350: Int = 350
 
 const val STEP_ONE: Int = 1
 const val STEP_TWO: Int = 2
