+++ app/src/main/java/com/example/comusenias/presentation/view_model/GenerateQRViewModel.kt
@@ -10,8 +10,6 @@ import com.example.comusenias.domain.models.users.ChildrenModel
 import com.example.comusenias.domain.repositories.QRRepository
 import com.example.comusenias.domain.use_cases.auth.AuthFactoryUseCases
 import com.example.comusenias.domain.use_cases.children.ChildrenFactory
-import com.example.comusenias.domain.use_cases.shared_preferences.DataUserStorageFactory
-import com.example.comusenias.domain.use_cases.specialist.SpecialistFactory
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers.IO
 import kotlinx.coroutines.launch
@@ -20,23 +18,20 @@ import javax.inject.Inject
 @HiltViewModel
 class GenerateQRViewModel @Inject constructor(
     private val childrenUseCases: ChildrenFactory,
-    private val specialistUseCases: SpecialistFactory,
     private val qrRepository: QRRepository,
     private val authUsesCases: AuthFactoryUseCases,
-    val dataStorageRepository: DataUserStorageFactory
 ) : ViewModel() {
 
     var state by mutableStateOf(QRState())
-    var userData by mutableStateOf(ChildrenModel())
+    private var userData by mutableStateOf(ChildrenModel())
         private set
-    val currentUser = authUsesCases.getCurrentUserUseCase()
-
+    private val currentUser = authUsesCases.getCurrentUserUseCase()
 
     init {
         startGenerate()
     }
 
-    fun startGenerate() = viewModelScope.launch(IO) {
+    private fun startGenerate() = viewModelScope.launch(IO) {
         currentUser?.let {
             childrenUseCases.getChildrenById(it.uid).collect { user ->
                 userData = user
